+++ src/Relude/Extra/Newtype.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP                 #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
 {- |
@@ -12,6 +13,10 @@ module Relude.Extra.Newtype
        ( un
        , wrap
        , under
+       , under2
+#if ( __GLASGOW_HASKELL__ != 802 )
+       , underF2
+#endif
        , (#.)
        ) where
 
@@ -19,6 +24,7 @@ import Relude
 
 -- $setup
 -- >>> :set -XTypeApplications
+-- >>> import Data.Semigroup (Max (..))
 
 {- | Unwraps value from @newtype@.
 
@@ -28,8 +34,9 @@ import Relude
 >>> un (Size 5) == length ['a', 'x', 'b']
 False
 -}
-un :: forall b a . Coercible a b => a -> b
+un :: forall a n . Coercible a n => n -> a
 un = coerce
+{-# INLINE un #-}
 
 {- | Wraps value to @newtype@. Behaves exactly as 'un' but has more meaningnful
 name in case you need to convert some value to @newtype@.
@@ -38,8 +45,9 @@ name in case you need to convert some value to @newtype@.
 >>> wrap False == Flag True
 False
 -}
-wrap :: forall b a . Coercible a b => a -> b
+wrap :: forall n a . Coercible a n => a -> n
 wrap = coerce
+{-# INLINE wrap #-}
 
 {- | Applies function to the content of @newtype@. This function is not supposed
 to be used on @newtype@s that are created with the help of smart constructors.
@@ -51,8 +59,35 @@ Foo False
 >>> under (filter (== 'a')) (Bar "abacaba")
 Bar "aaaa"
 -}
-under :: forall b a . Coercible a b => (b -> b) -> a -> a
+under :: forall n a . Coercible a n => (n -> n) -> (a -> a)
 under = coerce
+{-# INLINE under #-}
+
+{- | Lift binary function for @newtype@s to work over underlying @newtype@
+representation.
+
+>>> under2 @(Sum Int) (<>) (3 :: Int) 4
+7
+>>> under2 @All (<>) True False
+False
+-}
+under2 :: forall n a . Coercible a n => (n -> n -> n) -> (a -> a -> a)
+under2 = coerce
+{-# INLINE under2 #-}
+
+#if ( __GLASGOW_HASKELL__ != 802 )
+{- | Version of 'under2' that works on @newtype@s parametrized by their
+representation. Provided for convenience.
+
+>>> underF2 @Sum (<>) (3 :: Int) 4
+7
+>>> underF2 @Max (<>) 'p' 't'
+'t'
+-}
+underF2 :: forall n a . Coercible a (n a) => (n a -> n a -> n a) -> (a -> a -> a)
+underF2 = coerce
+{-# INLINE underF2 #-}
+#endif
 
 {- | Coercible composition
 -}