+++ src/Relude/Lifted.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE Safe #-}
-
 {- |
 Copyright: (c) 2016 Stephen Diehl
            (c) 20016-2018 Serokell
@@ -15,9 +13,11 @@ module Relude.Lifted
        , module Relude.Lifted.Exit
        , module Relude.Lifted.File
        , module Relude.Lifted.IORef
+       , module Relude.Lifted.Terminal
        ) where
 
 import Relude.Lifted.Concurrent
 import Relude.Lifted.Exit
 import Relude.Lifted.File
 import Relude.Lifted.IORef
+import Relude.Lifted.Terminal